+++ src/renderer/docs/propTypes/sortProps.ts
@@ -2,7 +2,7 @@ import type { PropDef } from 'storybook/internal/docs-tools';
 
 type Component = any;
 
-// react-docgen doesn't returned the props in the order they were defined in the "propTypes" object of the component.
+// docgen doesn't returned the props in the order they were defined in the "propTypes" object of the component.
 // This function re-order them by their original definition order.
 export function keepOriginalDefinitionOrder(
     extractedProps: PropDef[],