+++ src/renderer/entry-preview.ts
@@ -1,24 +1,31 @@
-/* Configuration for default renderer. */
-import { enhanceArgTypes } from 'storybook/internal/docs-tools';
+/* Configuration for default renderer (default preview.ts params) */
+import { global } from '@storybook/global';
+import { configure } from 'storybook/test';
 
-import { solidReactivityDecorator } from './renderToCanvas';
+import { solidReactivityDecorator } from './render';
 
 import type { Decorator } from './public-types';
-import type { ArgTypesEnhancer } from 'storybook/internal/types';
 
-export { render } from './render';
-export { renderToCanvas } from './renderToCanvas';
-export { applyDecorators } from './applyDecorators';
-export { mount } from './mount';
+export const parameters = {
+    renderer: 'solid',
+};
 
-export const parameters = { renderer: 'solid' };
-export const decorators: Decorator[] = [solidReactivityDecorator];
-export const argTypesEnhancers: ArgTypesEnhancer[] = [enhanceArgTypes];
+export const decorators: Decorator[] = [
+    solidReactivityDecorator,
+    (story, context) => {
+        // @ts-expect-error this feature flag not available in global storybook types
+        if (context.tags?.includes('test-fn') && !global.FEATURES?.experimentalTestSyntax) {
+            throw new Error(
+                'To use the experimental test function, you must enable the experimentalTestSyntax feature flag. See https://storybook.js.org/docs/10/api/main-config/main-config-features#experimentalTestSyntax'
+            );
+        }
+
+        return story();
+    },
+];
 
 export const beforeAll = async() => {
     try {
-        const { configure } = await import('storybook/test');
-
         configure({
             unstable_advanceTimersWrapper: (cb: () => any) => cb(),
             asyncWrapper: async(cb: () => any) => {
@@ -26,9 +33,11 @@ export const beforeAll = async() => {
 
                 await new Promise<void>((resolve) => {
                     setTimeout(() => resolve(), 0);
-                    if (jestFakeTimersAreEnabled()) {
+
+                    // @ts-expect-error global jest
+                    if (typeof jest !== 'undefined' && jest != null && ((setTimeout as any)._isMockFunction === true || Object.prototype.hasOwnProperty.call(setTimeout, 'clock'))) {
                         // @ts-expect-error global jest
-                        jest.advanceTimersByTime(0);
+                        jest!.advanceTimersByTime(0);
                     }
                 });
 
@@ -42,13 +51,6 @@ export const beforeAll = async() => {
     }
 };
 
-function jestFakeTimersAreEnabled() {
-    // @ts-expect-error global jest
-    if (typeof jest !== 'undefined' && jest !== null) {
-        // legacy timers or modern timers
-        return (setTimeout as any)._isMockFunction === true
-          || Object.prototype.hasOwnProperty.call(setTimeout, 'clock');
-    }
 
-    return false;
-}
+export { mount, render, renderToCanvas } from './render';
+export { applyDecorators } from './applyDecorators';