+++ src/renderer/index.ts
@@ -1,4 +1,4 @@
 import './globals';
 
 export * from './public-types';
-export * from './portable-stories';
+export { definePreview } from './preview';