+++ src/renderer/types.ts
@@ -2,12 +2,15 @@ import type { Component, JSXElement } from 'solid-js';
 
 import type { Args, Canvas, WebRenderer } from 'storybook/internal/types';
 
-export type { RenderContext, StoryContext } from 'storybook/internal/types';
+export type { RenderContext } from 'storybook/internal/types';
+export type { StoryContext } from './public-types';
+
+export type StoryFnReturnType = JSXElement;
 
 export interface SolidRenderer extends WebRenderer {
     // @ts-expect-error: Fix error in Github actions
     component: Component<this['T']>;
-    storyResult: StoryFnSolidReturnType;
+    storyResult: StoryFnReturnType;
     mount: (ui?: JSXElement) => Promise<Canvas>;
 }
 
@@ -16,12 +19,11 @@ export interface ShowErrorArgs {
     description: string;
 }
 
-export type ComponentsData = {
+export type GlobalReactivityStore = {
     [key: string]: {
         args: Args;
-        rendered?: boolean;
-        disposeFn?: () => void;
+        rendered: boolean;
+        disposeFn: (() => void) | null;
     };
 };
 
-export type StoryFnSolidReturnType = JSXElement;