+++ src/types.ts
@@ -1,15 +1,16 @@
 import type { BuilderOptions, StorybookConfigVite } from '@storybook/builder-vite';
+import type { ParserOptions } from 'react-docgen-typescript';
 import type {
     CompatibleString,
     StorybookConfig as StorybookConfigBase,
 } from 'storybook/internal/types';
 
-
-type FrameworkName = CompatibleString<'@storybook/html-vite'>;
+type FrameworkName = CompatibleString<'storybook-solidjs-vite'>;
 type BuilderName = CompatibleString<'@storybook/builder-vite'>;
 
 export type FrameworkOptions = {
     builder?: BuilderOptions;
+    docgen?: boolean | ParserOptions;
 };
 
 type StorybookConfigFramework = {
@@ -27,12 +28,15 @@ type StorybookConfigFramework = {
               options: BuilderOptions;
           };
     };
+    features?: StorybookConfigBase['features'] & {
+        /**
+         * Enable the experimental `.test` function in CSF Next
+         *
+         * @see https://storybook.js.org/docs/10/api/main-config/main-config-features#experimentalTestSyntax
+         */
+        experimentalTestSyntax?: boolean;
+    };
 };
 
 /** The interface for Storybook configuration in `main.ts` files. */
-export type StorybookConfig = Omit<
-    StorybookConfigBase,
-  keyof StorybookConfigVite | keyof StorybookConfigFramework
-> &
-StorybookConfigVite &
-StorybookConfigFramework;
+export type StorybookConfig = Omit<StorybookConfigBase, keyof StorybookConfigVite | keyof StorybookConfigFramework> & StorybookConfigVite & StorybookConfigFramework;