+++ tsup.config.ts
@@ -4,15 +4,19 @@ import { defineConfig } from 'tsup';
 export default defineConfig((options) => {
     return {
         entry: [
+            // framework
             'src/index.ts',
             'src/preset.ts',
+            'src/node/index.ts',
+            // renderer
             'src/renderer/index.ts',
+            'src/renderer/preview.ts',
             'src/renderer/preset.ts',
             'src/renderer/entry-preview.ts',
-            'src/renderer/entry-preview-docs.ts',
             'src/renderer/entry-preview-argtypes.ts',
+            'src/renderer/entry-preview-docs.ts',
         ],
-        format: ['esm', 'cjs'],
+        format: ['esm'],
         outDir: 'dist',
         dts: false,
         external: [